zebra: Track whether or not the dplane can use backup nexthops
Currently FRR has no ability to understand if the underlying
dplane has the ability to use backup nexthops.  Currently
in the linux kernel the answer is `no`.  But we have people
trying to use it as such.  Let's start the adding of breadcrumbs
so that end operators can figure this out.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>